Assume the default build target is 'all'
'all' is a well-known name

The default target must be explicitly named on the dry-run command line,
because of [Remaking Makefiles](https://www.gnu.org/software/make/manual/html_node/Remaking-Makefiles.html)
rule of GNU Make. In order to avoid remaking makefiles during a dry-run,
the makefile must also be given as an explicit target.

If the makefile were given as the only target, the dry-run would not
mean 'dry-run the default target', but 'dry-run nothing'. Thus, if the
makefile must be given as a target, the default target must be made
explicit, and thus is must be defined.

Also, previously 'all' was added to the make invokation for listing
targets. This would prevent the .DEFAULT_GOAL to be correctly set as the
Makefile declared, and instead be set to the value given on the command
line. An explicit target is no longer given on the --print-data-base
invocation, so the output .DEFAULT_GOAL will be the genuine default goal.

This commit makes currentTarget always defined, with the default value
'all'; this is a reasonable, well-known default.