Fix out-of-tree builds
out-of-tree builds allow multiple configurations to be built
simultaneously from one source tree. Each build configuration exists in
a separate builddir, and the srcdir contains no configuration, object
files or build artifacts. The builddir need not be a subdirectory of
srcdir, and typically it is on another filesystem than the sources.

In an autotools project which uses automake, the Makefiles are object
files, ie they are generated specifically for a particular
configuration, therefore they reside in the builddir.

When building an autotools project out-of-tree, the workspace is the
srcdir, and does not contain a Makefile. instead, the Makefile is in
$makeDirectory which is the builddir.

This commit cleans up the detection of makefile location that is done in
the extension so that the correct makefile is found, even if it lives
out-of-tree/out-of-workspace, has a name other than Makefile/makefile,
an whether or not a file named Makefile also exists in the srcdir.

The activationEvent is changed to "*" because when building out of
tree, there are no makefiles in the workspace. "*" is the better
activation because it means activate whenever the user enabled the
extension; don't second guess. OTOH, the extension does nothing
if a makefile is not found and not configured, so it's safe to
have it activated on workspaces without makefiles, because of the
cleaned-up detection/configuration of makefiles.