Avoid relativizing paths in the project outline
In the project outline view, paths for the Makefile and build log were
automatically made relative. If the setting was an absolute path, the
result would be ugly:

For a makeDirectory setting '/tmp/amhello-debug', the outline would print

Makefile: [../../../tmp/amhello-debug/Makefile]

This was not only unsightly, but also wrong. See
https://unix.stackexchange.com/questions/13858/do-the-parent-directorys-permissions-matter-when-accessing-a-subdirectory

In short, if one of the parent directories (..) would be inaccessible,
the ../../../tmp/amhello-debug path is wrong because it's inaccessible,
even though /tmp/amhello-debug is accessible.